Properly fix Premium+ Detection (#19)
* Properly fix Premium+ Detection

Move the Premium+ check to outside the Premium check (regex is looking
for exact match).

* Expand login fail check

Add status code 200 as a possible success response code.  200 is really
only applicible to Guest logins so we also check that too.  Logging in
normally still responds with 302 so we don't want to change that at all.